Move training_wrapper_functions to folder wrapper_functions
