+++ tests/test_training_wrapper_function.py
@@ -5,7 +5,7 @@ from matchms.exporting import save_as_mgf
 from matchms.importing import load_from_mgf
 from ms2deepscore.train_new_model.SettingMS2Deepscore import \
     SettingsMS2Deepscore
-from ms2deepscore.train_new_model.training_wrapper_functions import \
+from ms2deepscore.wrapper_functions.training_wrapper_functions import \
     train_ms2deepscore_wrapper
